Tweak recently-added tests to suppress scary-looking warnings on 64-bit machines about casts between pointers and integers of different sizes. While they're harmless, we shouldn't expect users to have to go through and figure that out for themselves.

+++ src/interfaces/ecpg/test/expected/thread-alloc.c
@@ -145,7 +145,7 @@ static void* fn(void* arg)
 #line 42 "alloc.pgc"
 
 
-	value = (int)arg;
+	value = (long)arg;
 	sprintf(name, "Connection: %d", value);
 
 	{ ECPGconnect(__LINE__, 0, "regress1" , NULL, NULL , name, 0); 
@@ -207,7 +207,7 @@ int main (int argc, char** argv)
 		CloseHandle(threads[i]);
 #else
 	for (i = 0; i < THREADS; ++i)
-		pthread_create(&threads[i], NULL, fn, (void*)i);
+		pthread_create(&threads[i], NULL, fn, (void *) (long) i);
 	for (i = 0; i < THREADS; ++i)
 		pthread_join(threads[i], NULL);
 #endif
